backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

langchain
langchain-community
langchain-openai
langchain-chroma
langchain_huggingface

chromadb
openai
python-dotenv==1.0.1
sentence-transformers
tiktoken==0.9.0
pypdf==4.2.0
pydantic
boto3==1.34.122
PyMuPDF==1.24.2
